Refactor customer reservation unit page
